requirements.txt
selenium==4.9.1
webdriver-manager==3.8.3
python-dotenv==1.0.0
requests==2.28.1
selenium-wire==5.1.0
